Added AndroidCamera project.
This is sample code that does several things:
- Uses android camera2 api to capture the back camera of the Android Device
- The camera stream is rendered to SurfaceTexture.
- OpenGLES 31 is initialized, and sample vertex and fragment shaders are provided.
- The camera stream is rendered on the device screen with initialised OpenGLES 31.

This sample code is intended to be starting point for other projects that will incorporate camera capture and real-time image processing - or rendering of the camera stream. Many fancy things can be done this way, like using the camera stream as a texture on 3D objects rendered by OpenGLES.

Android studio was used to develop this, so it is advisable to use it in order to be able to run this sample code just out of the box.